+++ publish/copyfiles.py
@@ -17,63 +17,19 @@ required folder structure:
 
 !!Note: anything excluded in .gitignore is not packaged by poetry 
 """
-import tomli_w
-from pathlib import Path
-import shutil
 import subprocess
-
-from create_project import create_project
+from pathlib import Path
 from typing import List, Tuple
+
+import tomli_w
 from packaging.version import parse
+
+from create_project import create_stub_package
 
 # todo : pass this as parameters
 ver = "v1_18"
 port = "esp32"
 board = "generic"
-
-
-def create_stub_package(
-    package_path: Path,
-    name: str,
-    stubs: List[Tuple[str, Path]],
-    version: str = "0.0.1",
-    description: str = "MicroPython stubs",
-):
-    """ """
-    # package_path = Path("C:\\develop\\MyPython\\micropython-stubs\\publish\\micropython-esp32-generic-stubs")
-
-    # todo: convert V1_18 to semver and use the patch level as version for the stubs package
-    semver = parse(version.replace("_", "."))
-
-    # create the package folder
-    package_path.mkdir(parents=True, exist_ok=True)
-
-    # Copy in the subs to the package
-    for name, folder in stubs:
-        print(f"Copying {name} from {folder}")
-        shutil.copytree(folder, package_path / folder.name, symlinks=True, dirs_exist_ok=True)
-
-    # add a readme with the names of the stubs
-    # todo: prettify this
-    with open(package_path / "README.md", "w") as f:
-        f.write(f"# {stub_package_name}\n\n")
-        f.write(f"Included stubs:\n")
-        for name, folder in stubs:
-            f.write(f"* {name} from {folder.as_posix()}\n")
-
-    # - create/update pyproject.toml
-    # todo: pass stubs
-    create_project(
-        package_path,
-        stub_package_name,
-        semver.public,
-        [p.name for description, p in stubs],
-        description=description,
-    )
-
-    # TODO: check the validity of the assembled package using mypy ?
-
-    subprocess.run(["poetry", "check"], cwd=package_path)
 
 
 # Defaults to the root of the project
